Create Multi-Signature-Wallet
After having spent the last weeks learning the ins and outs of the syntax in solidity, here is my first try at coding up a multi-sig wallet. It still needs further testing but for now I am quite satisfied with this. Smart Contract Engineer says that if you can code your own multi signature wallet, then you have a fairly good level in solidity. So this is a milestone for sure!